Changed memory type of owner property to assign, so that there are no retain cycles, Removed overwritten release, because dealloc already calls invalidate.
If you would e.g. call  [self.model removeAllBlockBasedObserversForOwner:self]; in your dealloc method,
it would never get called, because my controller (which is calling this) will never get deallocated!

I don't understand this comment: If this release will reduce the retain count to zero, prevent further calls to the block,
because, if this reduces the retain count to 0, dealloc will be called, which calls invalidate anyway